LibURL+LibWebView: Move public suffix data to LibURL
