+++ Libraries/LibURL/URL.cpp
@@ -1,6 +1,7 @@
 /*
  * Copyright (c) 2018-2020, Andreas Kling <andreas@ladybird.org>
  * Copyright (c) 2021, Max Wipfli <mail@maxwipfli.ch>
+ * Copyright (c) 2024, Sam Atkins <sam@ladybird.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -13,6 +14,10 @@
 #include <AK/Utf8View.h>
 #include <LibURL/Parser.h>
 #include <LibURL/URL.h>
+
+#if defined(ENABLE_PUBLIC_SUFFIX)
+#    include <LibURL/PublicSuffixData.h>
+#endif
 
 namespace URL {
 
@@ -496,6 +501,24 @@ ByteString percent_decode(StringView input)
         }
     }
     return builder.to_byte_string();
+}
+
+bool is_public_suffix([[maybe_unused]] StringView host)
+{
+#if defined(ENABLE_PUBLIC_SUFFIX)
+    return PublicSuffixData::the()->is_public_suffix(host);
+#else
+    return false;
+#endif
+}
+
+Optional<String> get_public_suffix([[maybe_unused]] StringView host)
+{
+#if defined(ENABLE_PUBLIC_SUFFIX)
+    return MUST(PublicSuffixData::the()->get_public_suffix(host));
+#else
+    return {};
+#endif
 }
 
 }
