+++ Libraries/LibURL/URL.h
@@ -199,6 +199,9 @@ URL create_with_url_or_path(ByteString const&);
 URL create_with_file_scheme(ByteString const& path, ByteString const& fragment = {}, ByteString const& hostname = {});
 URL create_with_data(StringView mime_type, StringView payload, bool is_base64 = false);
 
+bool is_public_suffix(StringView host);
+Optional<String> get_public_suffix(StringView host);
+
 }
 
 template<>
