+++ Libraries/LibWebView/URL.h
@@ -12,9 +12,6 @@
 
 namespace WebView {
 
-bool is_public_suffix(StringView host);
-Optional<String> get_public_suffix(StringView host);
-
 enum class AppendTLD {
     No,
     Yes,
