+++ Meta/Lagom/Tools/CodeGenerators/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(IPCCompiler)
 add_subdirectory(LibTextCodec)
+add_subdirectory(LibURL)
 add_subdirectory(LibWeb)
-add_subdirectory(LibWebView)
